convert the CollectionReportsController override into a decorator
